Move funcs to elements
